PLS-DA classification and feature selection

met.PLSDA.CV performs group classification and feature selection on a PLS model.

Usage

met.PLSDA.CV(
  mSetObj = NA,
  methodName = "LOOCV",
  k = 5,
  compNum = GetDefaultPLSCVComp(mSetObj),
  choice = "Q2",
  data = "all"
)

Arguments

mSetObj

Enter name of the created mSet object after PLS model creation (see met.PLSR.Anal).

methodName

(Character) Enter one of two methods for PLS-DA model (cross) validation:

"LOOCV" performs leave-one-out cross validation

"CV" performs k-fold cross validation.

k

(Numeric) The number of (randomized) groups that the dataset is to be split into during cross validation if methodName = "CV".

compNum

Number of components used for model validation (detected automatically).

choice

(Character) Choose the criterion used to estimate the predictive ability of the model, "Q2" or "R2".

data

(Character) Enter "all" to train the PLS(-DA) model on your whole (filtered and normalized) dataset or "anova" to use a subset of features defined as significant based on ANOVA analysis.

Value

The input mSet object with the results of PLS-DA at mSetObj$analSet$plsda.

References

adapted from PLSDA.CV (https://github.com/xia-lab/MetaboAnalystR).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com Technical University of Denmark License: GNU GPL (>= 2)