Perform T-test analysis

met.Ttests.Anal performs the Student's t-test on chosen sample groups. For large datasets (> 1000 variables), both the paired information and the group variance will be ignored, and the default parameters will be used for t-tests to save computational time. If you choose non-parametric tests (Wilcoxon rank-sum test), the group variance will be ignored.

Usage

met.Ttests.Anal(
  mSetObj = NA,
  grp1,
  grp2,
  nonpar = FALSE,
  threshp = 0.05,
  paired = FALSE,
  equal.var = TRUE,
  pvalType = "fdr",
  all_results = TRUE,
  silent = FALSE
)

Arguments

mSetObj

Enter the name of the created mSet object (see met.read_data).

grp1

(Character) Enter name of the first group for the contrast grp1 vs. grp2. If both group arguments are empty, the first two names in the list of groups are selected.

grp2

(Character) Enter name of the second group for the contrast grp1 vs. grp2. If both group arguments are empty, the first two names in the list of groups are selected.

nonpar

(Logical) Use a non-parametric test (TRUE) or not (FALSE).

threshp

(Numerical) Enter the adjusted p-value (FDR) cutoff.

paired

(Logical) Is the data paired (TRUE) or not (FALSE).

equal.var

(Logical) Is the group variance equal (TRUE) or not (FALSE).

pvalType

(Character) p value type used to apply significances based on the chosen threshold threshp. Can be "fdr" for adjusted p values or "raw" for raw p values.

all_results

(Logical) Create a CSV file with T-Test results for all compounds (TRUE) or not (FALSE).

silent

(Logical) Suppress message about the number of significant features found in the console (TRUE) or not (FALSE).

Value

The input mSet object with T-Test results added at mSetObj$analSet$tt$grp1_vs_grp2.

References

adapted from Ttests.Anal (https://github.com/xia-lab/MetaboAnalystR).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com Technical University of Denmark License: GNU GPL (>= 2)