Impute missing variables

Replace missing variables via a chosen method. Data need to be re-calibrated after this step, including met.PerformFeatureFilter as well as met.normalize. Data imputation is performed as part of the data preparation workflow met.read_data.

Usage

met.impute(mSetObj = NA, method = "min")

Arguments

mSetObj

Enter the name of the created mSet object ((see met.initialize and Read.TextData).

method

(Character) Select the option to replace missing variables:

"lod" replaces missing values with 1/5 of the minimum value for the respective variable.

"rowmin" replaces missing values with the half sample minimum.

"colmin" replaces missing values with the half feature minimum.

"mean" replaces missing values with the mean value of the respective feature column.

"median" replaces missing values with the median value of the respective feature column.

"knn_var" imputes missing values by finding the features in the training set “closest” to it and averages these nearby points to fill in the value.

"knn_smp" imputes missing values by finding the samples in the training set “closest” to it and averages these nearby points to fill in the value.

"bpca" applies Bayesian PCA to impute missing values.

"ppca" applies probabilistic PCA to impute missing values.

"svdImpute" applies singular value decomposition to impute missing values.

Value

The input mSet object with imputed data at mSetObj$dataSet$data_proc.

References

adapted from ImputeMissingVar (https://github.com/xia-lab/MetaboAnalystR).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com Technical University of Denmark License: GNU GPL (>= 2)