Data normalization

met.normalize performs row-wise normalization, transformation, and scaling of metabolomics data. This step is performed as part of the met.workflow function. Additionally, the workflow met.test_normalization allows the simultaneous testing of different data processing conditions and helps with finding the most suitable options.

Usage

met.normalize(
  mSetObj = NA,
  rowNorm = NULL,
  transNorm = NULL,
  scaleNorm = NULL,
  ref = NULL,
  norm.vec = NULL,
  ratio = FALSE,
  ratioNum = 20
)

Arguments

mSetObj

Enter the name of the created mSet object (see met.read_data).

rowNorm

(Character) Select the option for row-wise normalization:

"GroupPQN" for probabilistic quotient normalization by a reference group

"SamplePQN" for probabilistic quotient normalization by a reference sample

"QuantileNorm" for Quantile Normalization

"CompNorm" for normalization by a reference feature

"SumNorm" for normalization to constant sum of intensities

"MedianNorm" for normalization to sample median

"SpecNorm" for normalization by a sample-specific factor

transNorm

(Character) Select option to transform the data:

"LogNorm" for Log10 normalization

"CrNorm" Cubic Root Transformation

scaleNorm

(Character) Select option for scaling the data:

"MeanCenter" for Mean Centering

"AutoNorm" for Autoscaling

"ParetoNorm" for Pareto Scaling

"RangeNorm" for Range Scaling

ref

(Character) Enter the name of the reference sample or the reference feature (if rowNorm = "GroupPQN", "SamplePQN", or "CompNorm".

norm.vec

(Numeric vector) Vector with sample-specific scaling factors. Only applicable for rowNorm = "SpecNorm".

ratio

This option is only for biomarker analysis.

ratioNum

Relevant only for biomarker analysis.

Value

The input mSet object with normalized data at mSetObj$dataSet$norm.

References

adapted from Normalization (https://github.com/xia-lab/MetaboAnalystR).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com Technical University of Denmark License: GNU GPL (>= 2)