Visualize p values determined by ANOVA analysis.

met.plot_ANOVA visualizes the significance of each compound in the dataset.

Usage

met.plot_ANOVA(
  mSetObj = NA,
  imgName = "ANOVA_plot",
  format = "pdf",
  dpi = NULL,
  width = NA,
  subtitle = FALSE,
  export = TRUE,
  plot = TRUE
)

Arguments

mSetObj

Input name of the created mSet object, Data container after ANOVA analysis (met.ANOVA.Anal).

imgName

(Character) Enter a name for the image file (if export = TRUE).

format

(Character, "png" or "pdf") image file format (if export = TRUE).

dpi

(Numeric) resolution of the image file (if export = TRUE). If NULL, the resolution will be chosen automatically based on the chosen file format (300 dpi for PNG, 72 dpi for PDF)

width

(Numeric) width of the the image file in inches (if export = TRUE).

subtitle

(Logical, TRUE or FALSE) Shall the applied data transformation and scaling methods be displayed below the plot title?

export

(Logical, TRUE or FALSE) Shall the plot be exported as PDF or PNG file?

plot

(Logical, TRUE or FALSE) Shall the plot be returned in the RStudio 'Plots' pane?

Value

The input mSet object with added scatter plot with the p value for each compound analyzed. The plot can be retrieved from within R via print(mSetObj$imgSet$anova.plot). (generated by ggplot).

References

adapted from PlotANOVA (https://github.com/xia-lab/MetaboAnalystR).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com Technical University of Denmark License: GNU GPL (>= 2)