Plot PLS important features

met.plot_ImpVar is an internal function used by met.plot_PLS_Imp to visualize the features' importance in the PLS-DA model.

Usage

met.plot_ImpVar(
  mSetObj = NA,
  imp.vec,
  xlbl,
  feat.num = 15,
  color.BW = FALSE,
  title = FALSE
)

Arguments

mSetObj

Input name of the created mSet object, Data container after partial least squares-discriminant analysis (met.PLSR.Anal and met.PLSDA.CV).

imp.vec

A vector of importan variables.

xlbl

The x-label

feat.num

(Numeric) Indicate the number of features to show in the plot.

color.BW

(Logical) TRUE to use black and white, or FALSE to not.

title

(Logical) TRUE to add a title with the used normalization, transformation, and scaling method, or FALSE to not add any title.

Value

The input mSet object with added plot object.

References

adapted from PlotImpVar (https://github.com/xia-lab/MetaboAnalystR).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com