Plot a heatmap of compounds with missing values

met.plot_missval generates a heatmap of compounds with missing values to discover whether values are missing at random or not.

Usage

met.plot_missval(
  mSetObj,
  plot = TRUE,
  imgName = "Plots/MissValPlot",
  format = "pdf",
  export = TRUE,
  fontsize = 12
)

Arguments

mSetObj

Input name of the created mSet object, Data container with missing values.

plot

(Logical, TRUE or FALSE) Shall the plot be returned in the RStudio 'Plots' pane?

imgName

(Character) Enter a name for the image file (if export = TRUE).

format

(Character, "png" or "pdf") image file format (if export = TRUE).

export

(Logical, TRUE or FALSE) Shall the plot be exported as PDF and PNG file?

fontsize

(Numeric) Font sized used in the plot.

Value

The input mSet object with added heatmap indicating whether values are missing (0) or not (1). The plot can be retrieved from within R via ComplexHeatmap::draw(mSetObj$imgSet$missval_heatmap.plot) (generated by Heatmap).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com Technical University of Denmark License: GNU GPL (>= 2)