Generate a markdown report for the screening of metabolomics data pre-processing methods.

met.report_test_normalization generates a report of the analysis performed by the met.test_normalization wrapper function.

Usage

met.report_test_normalization(mSet_list, report.dir = NULL, ...)

Arguments

report.dir

(Character) Enter the name or location of the folder in which the report files are generated. If NULL (the default), a new folder "Report_date-time" is created in your working directory.

...

Further arguments passed from other methods.

mSetObj

Enter name of the created mSet object after a full analysis workflow (see met.workflow).

Value

In the specified folder, this function creates a PDF report, an HTML report, and an .RData file including the entire mSet object.

Author

Nicolas T. Wirth mail.nicowirth@gmail.com Technical University of Denmark License: GNU GPL (>= 2)