Pathway Enrichment Analysis

This function performs pathway enrichment analysis. Uses the DOSE package with the internal function enricher_internal.

Usage

pathway_enrich(
  gene,
  organism = "ppu",
  keyType = "kegg",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  use_internal_kegg = FALSE,
  custom_gene_sets = FALSE,
  custom_pathways = NULL
)

Arguments

gene

A numeric vector or a character vector of gene symbols that were found enriched in a differential expression analysis.

organism

A character string. The organism to use. Default is "ppu" for Pseudomonas putida KT2440.

keyType

A character string. The type of gene ID. Default is "kegg".

pvalueCutoff

A numeric value. The cutoff of p-value. Default is 0.05.

pAdjustMethod

A character string. The method of p-value adjustment. Default is "BH".

universe

A numeric vector or a character vector of gene symbols.

minGSSize

A numeric value. The minimum gene set size. Default is 10.

maxGSSize

A numeric value. The maximum gene set size. Default is 500.

qvalueCutoff

A numeric value. The cutoff of q-value. Default is 0.2.

use_internal_kegg

A logical value. Whether to use internal KEGG database. Default is FALSE.

custom_gene_sets

A logical value. Whether to use custom gene sets. Default is FALSE.

custom_pathways

A data frame. The custom pathways to use. Needs to have column "Pathway" and corresponding "Accession" column with comma-separated entries of gene/protein identifiers for that pathway.

Value

A data frame containing the results of the pathway enrichment analysis.