Impute missing values in a SummarizedExperiment object

Imputes missing values in a SummarizedExperiment object.

Usage

prot.impute(
  se,
  fun = c("bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "man", "min", "zero",
    "mixed", "nbavg", "SampMin"),
  ...
)

Arguments

se

A SummarizedExperiment object

fun

Character; imputation method, one of "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "man", "min", "zero", "mixed", "nbavg", or "SampMin".

...

Additional arguments passed to the imputation function.

Value

A SummarizedExperiment object with imputed values

Details

"SampMin" replaces missing values with the minimum value found in each sample. For information about the remaining imputation methods, see help("imputeMethods", "MsCoreUtils")

See also

Examples

data(mset)
#> Warning: data set ‘mset’ not found
se <- make_se(mset)
#> Error: object 'mset' not found
se <- prot.impute(se, fun = "knn")
#> Error: object 'se' not found