SCREE plot

prot.plot_screeplot() creates a SCREE plot of the explained variation in the principal components of a 'princomp' object

Usage

prot.plot_screeplot(
  pcaobj,
  components,
  xlim = NULL,
  ylim = c(0, 100),
  xlab = "Principal component",
  xlabAngle = 90,
  xlabhjust = 0.5,
  xlabvjust = 0.5,
  ylab = "Explained variation (%)",
  ylabAngle = 0,
  ylabhjust = 0.5,
  ylabvjust = 0.5,
  axisLabSize = 16,
  title = "SCREE plot",
  subtitle = "",
  caption = "",
  titleLabSize = 16,
  subtitleLabSize = 12,
  captionLabSize = 12,
  colBar = "dodgerblue",
  drawCumulativeSumLine = TRUE,
  colCumulativeSumLine = "red2",
  sizeCumulativeSumLine = 1.5,
  drawCumulativeSumPoints = TRUE,
  colCumulativeSumPoints = "red2",
  sizeCumulativeSumPoints = 2,
  hline = NULL,
  hlineType = "longdash",
  hlineCol = "black",
  hlineWidth = 0.4,
  vline = NULL,
  vlineType = "longdash",
  vlineCol = "black",
  vlineWidth = 0.4,
  gridlines.major = TRUE,
  gridlines.minor = TRUE,
  borderWidth = 0.8,
  borderColour = "black",
  plot = TRUE,
  export = TRUE
)

Arguments

pcaobj

An object of class 'princomp'

components

A numeric vector of principal component numbers to include in the plot

xlim

A numeric vector of length 2 specifying the x-axis limits

ylim

A numeric vector of length 2 specifying the y-axis limits. Default is c(0, 100)

xlab

A character string for x-axis label

xlabAngle

Numeric value for x-axis label angle

xlabhjust

Numeric value for x-axis label horizontal justification

xlabvjust

Numeric value for x-axis label vertical justification

ylab

A character string for y-axis label

ylabAngle

Numeric value for y-axis label angle

ylabhjust

Numeric value for y-axis label horizontal justification

ylabvjust

Numeric value for y-axis label vertical justification

axisLabSize

Numeric value for axis label size

title

A character string for the plot title

subtitle

A character string for the plot subtitle

caption

A character string for the plot caption

titleLabSize

Numeric value for title label size

subtitleLabSize

Numeric value for subtitle label size

captionLabSize

Numeric value for caption label size

colBar

A character string for the color of the bar chart

drawCumulativeSumLine

Logical value to indicate if the cumulative sum line should be plotted

colCumulativeSumLine

A character string for the color of the cumulative sum line

sizeCumulativeSumLine

Numeric value for the size of the cumulative sum line

drawCumulativeSumPoints

Logical value to indicate if the cumulative sum points should be plotted

colCumulativeSumPoints

A character string for the color of the cumulative sum points

sizeCumulativeSumPoints

Numeric value for the size of the cumulative sum points

hline

A numeric value for horizontal line

hlineType

A character string for the type of horizontal line

hlineCol

A character string for the color of the horizontal line

hlineWidth

Numeric value for the width of the horizontal line

vline

A numeric value for vertical line

vlineType

A character string for the type of vertical line

vlineCol

A character string for the color of the vertical line

vlineWidth

Numeric value for the width of the vertical line

gridlines.major

Logical value to indicate if major gridlines should be plotted

gridlines.minor

Logical value to indicate if minor gridlines should be plotted

borderWidth

Numeric value for the width of the border

borderColour

A character string for the color of the border

plot

Logical value to indicate if the plot should be plotted

export

Logical value to indicate if the plot should be exported as PNG and PDF

Value

A SCREE plot