Title

Title

Usage

prot_to_map(
  se,
  protein_set = NULL,
  col.id = "ID",
  select.id = "Accession",
  pfx = NULL,
  imp_fun = c("man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min", "zero",
    "mixed", "nbavg", "SampMin"),
  q = 0.01,
  type = c("contrast", "centered"),
  condition = NULL,
  contrast = NULL,
  template = NULL,
  RSD_thresh = 0.5,
  p_thresh = 1,
  indicate_nonsig = c("gray", "dashed"),
  result = NULL,
  pal = "RdYlBu",
  legend_min = NULL,
  legend_max = NULL,
  title = "",
  export = TRUE,
  inkscape_path = "C:/Program Files/Inkscape/bin/inkscape.exe",
  export_dpi = 300,
  export_width = 2281,
  export_height = 2166
)

Arguments

se

SummarizedExperiment object, proteomics data parsed with prot.read_data.

protein_set

Data frame containing a list of proteins to be plotted. The provided names in column must match the entries in the column provided as col.id as well as the IDs of objects in the map template.

col.id

(Character) Column name of the protein identifier in the SummarizedExperiment object.

select.id

Column name of the protein identifier in the protein_set data frame.

pfx

Prefix of the protein identifiers to be plotted. all protein objects in the template map should have a common prefix (e.g., 'prot_')

imp_fun

(Character string) Function used for data imputation. "SampMin", "man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min", "zero", "mixed", or "nbavg". See (prot.impute) for details.

q

(Numeric) q value for imputing missing values with method imp_fun = 'MinProb'.

type

(Character string) Type of analysis to perform. "contrast" (provide a specific contrast in the contrast argument to plot log2-fold change values) or "centered" (provide a single condition as condition to plot centered abundances in that condition).

condition

Condition to be displayed if type = "centered".

contrast

Tested contrast if type = "contrast" in the format "A_vs_B".

template

Path to the SVG map template.

p_thresh

(adjusted) p-value threshold.

result

Name of the output files.

pal

Brewer color palette used for protein squares.

legend_min

Lower limit of the color legend.

legend_max

Upper limit of the color legend.

title

(Character string) Title shown above the map

export

(Logical) A logical value indicating whether to export the plot as a PNG and PDF file.

inkscape_path

The local path to the 'inkscape.exe' file. Required if export = TRUE.

export_dpi

The dpi of the exported PNG and PDF file.

export_width

The width of the exported PNG and PDF file. The width of the template SVG can be inspected after opening it in InkScape.

export_height

The height of the exported PNG and PDF file. The height of the template SVG can be inspected after opening it in InkScape.

RSD_threshRelative

standard deviation threshold.

Value

A metabolic map with protein objects colored according to their abundance or log2 fold changes in a given contrast.