Plot Results of Principal Component Analysis

Performs 'regularized log' transformation followed by PCA on a given DESeqDataSet object and plots the results.

Usage

rna.plot_pca(
  dds,
  x = 1,
  y = 2,
  indicate = c("condition", "replicate"),
  title = NULL,
  label = FALSE,
  n = ncol(t(SummarizedExperiment::assay(dds))),
  point_size = 4,
  label_size = 3,
  hline = 0,
  hlineType = "longdash",
  hlineCol = "black",
  hlineWidth = 0.4,
  vline = 0,
  vlineType = "longdash",
  vlineCol = "black",
  vlineWidth = 0.4,
  basesize = 15,
  plot = TRUE,
  export = TRUE
)

Arguments

x

x-axis PC (default: 1)

y

y-axis PC (default: 2)

indicate

features to indicate in the plot ("condition" or "replicate")

title

title of the plot (default: "PCA plot - top n variable genes")

label

whether to label points in the plot (default: FALSE)

n

number of variables to take into account, sorted according to their variance in descending order. Only the n most variable genes are used to perform PCA. default: number of columns in dep)

point_size

size of points in the plot (default: 4)

label_size

size of labels in the plot (default: 3)

hline

position of horizontal line (default: 0)

hlineType

type of horizontal line (default: 'longdash')

hlineCol

color of horizontal line (default: 'black')

hlineWidth

width of horizontal line (default: 0.4)

vline

position of vertical line (default: 0)

vlineType

type of vertical line (default: 'longdash')

vlineCol

color of vertical line (default: 'black')

vlineWidth

width of vertical line (default: 0.4)

basesize

base size of the plot (default: 15)

plot

whether to return the PCA plot (default: TRUE)

export

whether to export the PCA plot to the Plots directory as PNG and PDF file (default: TRUE)

dep

DESeqDataSet object

Value

(invisibly) a data frame containing the PCA coordinates