Plots a volcano plot for a given contrast

Plots a volcano plot for a given contrast

Usage

rna.plot_volcano(
  dds,
  contrast,
  label_size = 3,
  alpha = 0.05,
  lfc = 2,
  x.lim = NULL,
  add_names = TRUE,
  adjusted = FALSE,
  plot = TRUE,
  export = TRUE
)

Arguments

dds

A DESeqDataSet object

contrast

The contrast of interest in the format "conditionA_vs_conditionB"

label_size

The size of labels for the gene names

alpha

The alpha value used to determine significance.

lfc

The log2 fold change threshold.

add_names

Logical, whether to add gene names to the plot

adjusted

Logical, whether to plot adjusted p-values. The alpha threshold is applied to adjusted p values even if adjusted = FALSE.

plot

Logical, whether to return the volcano plot

export

Logical, whether to export the volcano plot as PND and PDF file

Value

(invisibly) A data frame with columns for gene, log2 fold change, and -log10 p-value