Convert SVG to PNG (requires Python and InkScape installed)

This function converts an SVG file to a PNG file by calling the respective InkScape function through Python.

Usage

svg_to_png(
  svg = svg,
  inkscape_path = "C:/Program Files/Inkscape/bin/inkscape.exe",
  out = NULL,
  dpi = 300,
  width = 2281,
  height = 2166
)

Arguments

svg

A character string specifying the path of the SVG file to be converted.

inkscape_path

The local path to the 'inkscape.exe' file.

out

A character string specifying the desired name of the output file. If NULL (the default), the output file will have the same name as the input file, but with a .png extension.

dpi

A numeric value specifying the desired dots per inch of the output file. The default is 300.

width

A numeric value specifying the desired width of the output file. The default is 2281.

height

A numeric value specifying the desired height of the output file. The default is 2166.

Value

A message specifying the location of the output file.