Visualize effect of normalization on sample density distributions.

met.plot_SampleNormSummary visualizes the density distributions of samples before and after normalization, transformation, and scaling.

Usage

met.plot_SampleNormSummary(
  mSetObj = NA,
  imgName = "SampleNormSummary",
  format = "png",
  dpi = NULL,
  width = NA,
  show_prenorm = TRUE,
  export = TRUE,
  plot = TRUE
)

Arguments

mSetObj

Input name of the created mSet object, Data container after normalization (met.normalize).

imgName

(Character) Enter a name for the image file (if export = TRUE).

format

(Character, "png" or "pdf") image file format (if export = TRUE).

dpi

(Numeric) resolution of the image file (if export = TRUE). If NULL, the resolution will be chosen automatically based on the chosen file format (300 dpi for PNG, 72 dpi for PDF)

width

(Numeric) width of the the image file in inches (if export = TRUE).

show_prenorm

(Logical, TRUE or FALSE) Shall density distributions before normalization be displayed?

export

(Logical, TRUE or FALSE) Shall the plot be exported as PDF or PNG file?

plot

(Logical, TRUE or FALSE) Shall the plot be returned in the RStudio 'Plots' pane?

Value

The input mSet object with added plot: the top is a density plot and the bottom is a box plot. In a boxplot, the bottom and top of the box are always the 25th and 75th percentile (the lower and upper quartiles, or Q1 and Q3, respectively), and the band near the middle of the box is always the 50th percentile (the median or Q2). The upper whisker is located at the smaller of the maximum x value and Q3 + 1.5 x IQR (Interquantile Range), whereas the lower whisker is located at the larger of the smallest x value and Q1 - 1.5 x IQR. The plot can be retrieved from within R via print(mSetObj$imgSet$summary_norm_sample.plot).

References

adapted from PlotSampleNormSummary (https://github.com/xia-lab/MetaboAnalystR).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com