Create heat map of hierarchically clustered features and samples

met.plot_heatmap plots a heat map with optional filtering for top features based on results from t-tests/ANOVA, VIP or randomforest.

Usage

met.plot_heatmap(
  mSetObj = NA,
  imgName = "Heatmap_features",
  format = "pdf",
  dpi = NULL,
  width = NA,
  dataOpt = "norm",
  scaleOpt = "row",
  smplDist = "euclidean",
  clstDist = "ward.D",
  palette = "bwm",
  viewOpt = "detail",
  rowV = T,
  colV = T,
  var.inx = NULL,
  border = T,
  grp.ave = F,
  plot = TRUE,
  export = TRUE
)

Arguments

mSetObj

Input name of the created mSet object.

imgName

(Character) Enter a name for the image file (if export = TRUE).

format

("png" or "pdf") image file format (if export = TRUE).

dpi

(Numeric) resolution of the image file (if export = TRUE). If NULL, the resolution will be chosen automatically based on the chosen file format (300 dpi for PNG, 72 dpi for PDF)

width

(Numeric) width of the the image file in inches (if export = TRUE).

dataOpt

(Character) Set data options. Choose "norm" to use data after normalization or any other string to use original values.

scaleOpt

(Character) Set value standardization option. "row" for autoscaling of features, "column" for autoscaling of samples, "none" for no autoscaling.

smplDist

(Character) Enter the sample distance method: "euclidean" for Euclidean distance, "correlation" for Pearson distance, or "minkowski" for Minkowski method.

clstDist

(Character) Enter the feature clustering method: "ward.d", "average", "complete", or "single".

palette

Input color palette choice: "bwm" for Blue-White-Red, "gbr" for Green-Black-Red, "heat" for Red-Yellow, "topo" for topological colors, "gray" for gray-scale.

viewOpt

Set view options, choose "detail" or "overview".

rowV

(Logical) Shall clustering be applied to samples?

colV

(Logical) Shall clustering be applied to features?

var.inx

(Numeric vector) optional: Provide indeces of selected features to be plotted.

border

(Logical) Indicate whether or not to show cell-borders.

grp.ave

(Logical) Shall only group averages be shown (as opposed to displaying every sample)?

plot

(Logical, TRUE or FALSE) Shall the plot be returned in the RStudio 'Plots' pane?

export

(Logical, TRUE or FALSE) Shall the plot be exported as PDF or PNG file?

Value

The input mSet object with added heat map. The plot can be retrieved from within R via print(mSetObj$imgSet$heatmap.avg.plot) or print(mSetObj$imgSet$heatmap.all.plot), depending on grp.ave.

References

adapted from PlotHeatMap (https://github.com/xia-lab/MetaboAnalystR).

Author

Nicolas T. Wirth mail.nicowirth@gmail.com