PCA Analysis

This function performs principal component analysis (PCA) on a given matrix.

Usage

prot.pca(
  mat,
  metadata = NULL,
  center = TRUE,
  scale = FALSE,
  rank = NULL,
  removeVar = NULL,
  transposed = FALSE,
  BSPARAM = ExactParam(),
  in_workflow = FALSE
)

Arguments

mat

A numeric matrix.

metadata

An optional data frame with rownames matching 'colnames(mat)'.

center

A logical indicating whether to center the data before PCA.

scale

A logical indicating whether to scale the data before PCA.

rank

An integer indicating the number of principal components to calculate.

removeVar

A numeric value between 0 and 1 indicating the fraction of variables to remove based on variance.

transposed

A logical indicating whether the matrix is transposed.

BSPARAM

An object of class ExactParam or ApproxParam.

in_workflow

Indicated whether this function is called from within prot.workflow.

Value

A list with components rotated, loadings, variance, sdev, metadata, xvars, yvars, and components.

References

BiocSingular (https://bioconductor.org/packages/release/bioc/html/BiocSingular.html)

Examples

mat <- matrix(rnorm(1000), nrow = 100, ncol = 10)
prot.pca(mat)
#> $rotated
#>           PC1        PC2        PC3        PC4        PC5        PC6        PC7
#> 1  -0.2181752  0.4813187 -2.7276185  1.7155016 -6.4659811  1.8272685  0.4353507
#> 2   0.9613639 -4.8789898  4.0097033 -1.2816326  1.7080328 -4.8053296 -0.4818599
#> 3   7.9541566  0.4110740  2.1803831 -0.4518610 -0.4823244  0.4668927  4.5767488
#> 4  -1.4502359  4.8306645  6.1535514 -3.8142828 -2.5934762  1.3263417 -2.3976805
#> 5  -3.8871078  2.7582896 -1.2124923 -2.7523709  5.6949803  2.8408021  1.3133760
#> 6   2.5763750  3.6725005 -0.3941912  6.1745695  1.9981151 -1.7445527 -4.2183221
#> 7   1.6470210 -5.0552184  0.1094468  2.4699898  1.9730844  5.3829079 -1.0074009
#> 8  -8.1826658 -2.5254038  1.6717288  3.3639424 -1.4240932 -1.2617544  2.3802975
#> 9   0.6657211 -3.4510675 -5.0781177 -5.2451324 -1.2538593 -0.9465269 -3.0777605
#> 10 -0.0664530  3.7568322 -4.7123936 -0.1787236  0.8455215 -3.0860491  2.4772509
#>          PC8        PC9         PC10
#> 1  -1.190265  4.1223588 7.118897e-16
#> 2  -2.299750  2.8843376 7.118897e-16
#> 3   2.958072 -0.6562997 7.118897e-16
#> 4  -1.136323 -1.7435364 7.118897e-16
#> 5   1.253705  3.1191257 7.118897e-16
#> 6   2.007674  0.3061732 7.118897e-16
#> 7  -2.896258 -2.0519892 7.118897e-16
#> 8   2.558773 -1.9975495 7.118897e-16
#> 9   2.867369 -1.4701931 7.118897e-16
#> 10 -4.122996 -2.5124275 7.118897e-16
#> 
#> $loadings
#>              PC1          PC2          PC3           PC4          PC5
#> 1    0.043417958  0.108486710 -0.083956007  7.805625e-02  0.065641809
#> 2    0.070830340 -0.056583940  0.045929652 -9.897265e-02  0.018809116
#> 3    0.021638669  0.038243056  0.071424967  1.688323e-01 -0.031031497
#> 4   -0.002967440  0.020880334  0.079959036  7.612662e-02  0.009471666
#> 5    0.215354907 -0.181537187  0.055165171 -6.662039e-02  0.078252200
#> 6    0.036630863 -0.051531794 -0.010495424 -5.582335e-02  0.211709675
#> 7   -0.161633310 -0.145109288  0.052381040  2.007621e-01 -0.002457323
#> 8    0.095294386 -0.119624281 -0.009573547  2.052435e-02 -0.026147537
#> 9   -0.121458359  0.180683534  0.062845707  1.673614e-01  0.092873490
#> 10   0.011042698  0.012063392  0.098758164 -2.242717e-02 -0.136171012
#> 11  -0.170591502  0.009062904  0.073277007  6.721721e-02  0.047193216
#> 12   0.086746689  0.054850626 -0.024925725 -7.873895e-02 -0.040361291
#> 13  -0.086900844  0.194406016  0.104171284  1.044268e-01 -0.054175400
#> 14   0.059427649  0.112827319  0.105725885  4.855546e-02  0.161735510
#> 15   0.015350124  0.104656312 -0.046949584 -5.685033e-02  0.018100069
#> 16  -0.080448728  0.241317341  0.186918052 -1.453082e-01 -0.075757252
#> 17   0.108073188 -0.104205601  0.071008461  6.988622e-02  0.063459315
#> 18   0.098696155  0.084905078 -0.095186570  5.264372e-02 -0.137576008
#> 19  -0.067461855 -0.123549634 -0.226648503 -3.471837e-02  0.026906248
#> 20  -0.034845040  0.001266117  0.070822364  8.000786e-04  0.127962575
#> 21  -0.011782449  0.089826475  0.148371317 -1.594786e-01  0.092474113
#> 22   0.064937890 -0.005197635 -0.165044557 -5.760863e-02  0.079693517
#> 23   0.012587406  0.040405307  0.094795801  1.893187e-01 -0.033264027
#> 24   0.081918057 -0.297046024  0.090740965 -6.031396e-02 -0.074921501
#> 25   0.237752570  0.147262297  0.080053543 -9.019533e-02 -0.106010448
#> 26  -0.077224665  0.098576683  0.010691453 -6.522864e-02  0.147277826
#> 27   0.182902620  0.009831160  0.127701806  9.444600e-02  0.048393033
#> 28   0.108951303  0.173693923  0.012541495 -1.438892e-01 -0.042302048
#> 29   0.010342471  0.189833593 -0.014220176  1.369693e-01 -0.103305982
#> 30   0.030901128 -0.122056456 -0.106559788  2.020167e-02 -0.059564774
#> 31   0.016669886  0.120399954 -0.244132708 -4.760750e-02 -0.028085240
#> 32   0.002594391 -0.081803213  0.140367435 -1.073044e-01 -0.019537676
#> 33   0.325974782 -0.052867138  0.155370349 -9.204286e-02 -0.030798056
#> 34  -0.176595823  0.025753266 -0.061042508  1.234154e-01  0.031013558
#> 35  -0.056723268 -0.037847380  0.078225339  1.710807e-02 -0.154223026
#> 36  -0.003309639  0.013312982  0.094024595 -1.290950e-01  0.097848153
#> 37   0.139894185 -0.006232768  0.073164272 -2.008029e-01 -0.079664524
#> 38  -0.001451855 -0.033559010 -0.170126963 -7.460242e-02  0.085527997
#> 39   0.087152250  0.196742141 -0.151855571 -9.405637e-03 -0.121485249
#> 40   0.018104641  0.171202611  0.107283035 -1.165696e-01  0.007090749
#> 41   0.093908656  0.061048269 -0.122096947 -4.509223e-02  0.135485052
#> 42   0.071587601 -0.061510510  0.142348259  5.312614e-02  0.070040201
#> 43   0.008939514  0.100752711  0.099682055 -3.420179e-02  0.145351725
#> 44  -0.060131352 -0.030244612  0.121695120  6.330368e-02  0.006600888
#> 45   0.008704579  0.081081512 -0.080451492  1.779350e-01  0.135883469
#> 46   0.067134039  0.039086031  0.023908950  1.214696e-01  0.070130341
#> 47  -0.023827842  0.027287032  0.190423814  9.962561e-02 -0.250014892
#> 48  -0.040136847  0.075332602  0.038588366  1.947538e-02  0.125729774
#> 49   0.174339103 -0.096846328 -0.156973922 -3.090479e-02  0.169311981
#> 50  -0.029150204 -0.024786414  0.009727860 -6.724564e-05 -0.171861931
#> 51  -0.050563785  0.008632527  0.052895801  1.635029e-02  0.088662141
#> 52   0.084370537 -0.095046418 -0.021390841 -3.643207e-02  0.079979343
#> 53   0.105225554 -0.115803804  0.131748214  2.337456e-01  0.039534603
#> 54   0.167767464  0.032402905  0.021069833 -2.216250e-03 -0.053378660
#> 55   0.179238330  0.126982377 -0.006744840  1.636447e-03  0.104311757
#> 56  -0.041348565  0.128127933 -0.063006942 -1.836588e-01 -0.017902969
#> 57  -0.028230886  0.021196776 -0.068166393 -5.322601e-03  0.059965938
#> 58   0.040906329 -0.077579930  0.052971013  9.206470e-03 -0.149682660
#> 59   0.013093440 -0.030160112 -0.059945307 -1.507064e-02 -0.094915092
#> 60  -0.099136311 -0.006380889  0.131620018 -4.625063e-02  0.186929411
#> 61   0.134234582 -0.009455837 -0.147997644  2.196880e-02 -0.017630017
#> 62   0.007249927  0.026158426 -0.040677006  1.890066e-01  0.086056413
#> 63   0.067397044 -0.054542113  0.053586434  2.395037e-02  0.004824324
#> 64   0.018247550  0.059694290 -0.025716766  2.046027e-01  0.016394446
#> 65   0.133032057  0.171419799 -0.069556116  7.000782e-02  0.187225394
#> 66  -0.000800276 -0.039977547  0.073720279  1.076931e-02 -0.005506736
#> 67   0.032862641  0.041540821  0.008829407  8.801194e-02  0.064361600
#> 68  -0.153051910 -0.037691818  0.180004305 -7.700821e-02 -0.053699778
#> 69  -0.104296675 -0.090424554 -0.067013455  4.101589e-02 -0.105121045
#> 70  -0.126351262 -0.012724835  0.045937048 -1.102126e-01 -0.008722538
#> 71  -0.044653354  0.019936259 -0.001471407  4.029317e-02 -0.036225789
#> 72   0.117321626  0.084531550 -0.002509632  1.275527e-01 -0.023776657
#> 73  -0.044327641  0.047732796  0.290420415  5.762729e-02  0.070080660
#> 74  -0.088934757 -0.023715420 -0.006865677  1.262470e-01 -0.013210713
#> 75   0.147081085  0.106662206  0.022616303  4.894814e-02  0.173386474
#> 76  -0.153854511 -0.055365916 -0.145030994 -1.158742e-01  0.238436334
#> 77  -0.015177884  0.066607271 -0.196096475 -1.878019e-02 -0.083367076
#> 78   0.349318026 -0.175750757 -0.025943617  1.753233e-01 -0.117636441
#> 79  -0.059853063 -0.191099311  0.086394293 -1.344038e-01  0.079311173
#> 80  -0.079368598  0.047453569  0.023925378  2.326213e-01  0.170673356
#> 81   0.166511929  0.123876663  0.046882578  3.180892e-02 -0.048111842
#> 82  -0.007765441 -0.048779412  0.034923977 -7.145869e-02  0.064764757
#> 83  -0.003010083  0.018164681 -0.133273390 -1.275960e-01  0.059965893
#> 84  -0.014519185  0.115134951  0.058437296  4.528436e-02 -0.122707350
#> 85  -0.064208550 -0.009354248 -0.025239872 -1.351939e-02 -0.125987689
#> 86  -0.108437601 -0.118647213 -0.144061670  5.275766e-02 -0.023539597
#> 87   0.040658109  0.131064316  0.109512893 -1.418767e-01  0.003366520
#> 88  -0.052558272 -0.057201671  0.036152349 -1.277879e-01  0.036199234
#> 89  -0.020827700 -0.172060597  0.129911790 -6.910519e-02  0.004619733
#> 90  -0.108336794 -0.069997522  0.098068941 -5.636000e-03  0.002571307
#> 91   0.012643366  0.077731808  0.058981230  1.088388e-02  0.099315486
#> 92  -0.083485297 -0.001172631 -0.107309926  4.002336e-02 -0.240666648
#> 93   0.005664383 -0.026973344 -0.020928639  4.599197e-02  0.051619570
#> 94   0.038293773 -0.082784967 -0.030909043  9.244753e-02  0.109004986
#> 95   0.063861010 -0.031816853  0.063106037  1.028120e-01  0.076429528
#> 96   0.038833877  0.007982063  0.014386390 -5.812568e-02  0.112572655
#> 97   0.038562283  0.014533369 -0.082723125  1.911047e-01 -0.067027612
#> 98  -0.041703301  0.046805281 -0.012569403 -5.738930e-02  0.072708133
#> 99   0.026832308 -0.164675103  0.072677724  6.253273e-02  0.159659943
#> 100 -0.010952137 -0.131323703 -0.030135310  7.814111e-02  0.063703342
#>              PC6          PC7           PC8           PC9          PC10
#> 1   -0.027465489 -0.126206971 -0.1539099525  0.0465643461  0.2935832219
#> 2    0.152248620 -0.105680541 -0.0909311002  0.0969767547  0.0505896005
#> 3   -0.116876707 -0.051547717  0.1544360242 -0.0204759651  0.0047791050
#> 4    0.214623813  0.177654858  0.0719352327 -0.0766427540  0.0702629422
#> 5   -0.215461506 -0.124183266 -0.0315353553  0.0896451161  0.3867309256
#> 6   -0.003780699  0.006893998 -0.0686369194  0.1397579215 -0.1661019886
#> 7   -0.073884497 -0.106762628 -0.1153023562 -0.0726223718 -0.6355610604
#> 8    0.211810828  0.006942882  0.0708383181  0.0428208598  0.1289283205
#> 9    0.048203869  0.024408928 -0.0677643676  0.0313582917 -0.0474081529
#> 10  -0.012845592  0.151687343 -0.0173166648 -0.0259853806 -0.0150518020
#> 11   0.037851588  0.016212018  0.1094924940  0.0550375136 -0.0550376203
#> 12   0.053842783 -0.056434377  0.0131757810 -0.1046508047 -0.0383653545
#> 13   0.061316483  0.205111351 -0.1459880195  0.0342895075  0.1104572498
#> 14  -0.068350667 -0.001185560  0.0583211898 -0.0290879781  0.0608654851
#> 15  -0.066356942  0.113077650  0.0167505601  0.0878041830 -0.0183298214
#> 16  -0.133587916 -0.046144518 -0.0343073475 -0.0034638879 -0.0497221435
#> 17  -0.022884226  0.113630452  0.0113158760 -0.0380432959 -0.0039902298
#> 18  -0.060039488 -0.060218594 -0.3404975480 -0.0723699473 -0.0558630969
#> 19   0.044844761 -0.008815307 -0.0858746070 -0.0681082307  0.0141956964
#> 20   0.143580739 -0.054966317 -0.0079104434 -0.0212508489 -0.0699925396
#> 21   0.005632094 -0.139689369 -0.0637808811 -0.0578830803  0.0048324640
#> 22  -0.016551880 -0.020717851 -0.0660279610  0.0266922658 -0.0209204409
#> 23  -0.077691461 -0.094236928  0.0732906426 -0.0752829967 -0.0100405910
#> 24  -0.045093645  0.071056405  0.0010565386  0.0839556396  0.0584762530
#> 25   0.114211578 -0.013498988 -0.0218358599 -0.1466342332 -0.0276867440
#> 26  -0.117350995  0.135844991 -0.0962018285  0.0500654858  0.0723867529
#> 27  -0.022623978  0.036023627 -0.1728307985  0.2139999786 -0.0372151334
#> 28   0.164815491 -0.038056659  0.0614411311  0.2502467043 -0.1431852070
#> 29  -0.209295729  0.041554003  0.1607115210  0.0127377750 -0.0098064292
#> 30   0.062423307 -0.121923760 -0.1224180721 -0.1225797374  0.0977356687
#> 31  -0.144915680  0.063651340  0.0269543677  0.2221688780 -0.0216789118
#> 32  -0.055003869 -0.066554993  0.0674232751  0.0558009963  0.0072580766
#> 33  -0.001544347 -0.159971052  0.1607004534 -0.2288935791 -0.0398647793
#> 34  -0.050316873 -0.051675020 -0.0265304310 -0.1514474505  0.1274425365
#> 35   0.046442655 -0.053458776 -0.0369720039 -0.1549399339  0.0686346274
#> 36  -0.006011091  0.002417284 -0.0352611642  0.0192566888 -0.0387909502
#> 37  -0.113903903 -0.056998764  0.1082263141  0.0453958804 -0.0549145358
#> 38   0.023725416 -0.053835999 -0.0023382824 -0.1307866476 -0.1281994193
#> 39   0.144858609  0.084905330  0.0501695349 -0.0136627614 -0.0024649223
#> 40  -0.007066479 -0.055354973 -0.2493176313  0.0459525421  0.0634605008
#> 41   0.071096538  0.028560201 -0.0076901696  0.0045809451 -0.1091644175
#> 42   0.048258748  0.022243061 -0.1080823115  0.0373357293 -0.0110499642
#> 43   0.033822418 -0.058748988  0.0619653412 -0.0172320646 -0.0169490125
#> 44   0.186054350  0.161190551 -0.0432176321  0.1355353752  0.1494527782
#> 45   0.082597609 -0.113694194  0.0750162181  0.1752833141 -0.0662250909
#> 46   0.038787476  0.004259795 -0.0017201241 -0.0417476375  0.0464107834
#> 47  -0.300880147  0.079645680  0.0564990310  0.0033140372 -0.0804519458
#> 48   0.053654849 -0.112382882  0.0106975663 -0.0055610878  0.0364681615
#> 49  -0.218887840 -0.019334991  0.0595928050 -0.0881910369 -0.0119424254
#> 50   0.116068800 -0.084628718  0.1485391619  0.0119900964 -0.0461713262
#> 51   0.013030551 -0.017601952  0.0925109676 -0.0785631206  0.0662755733
#> 52   0.074572840  0.233102097 -0.0743325949 -0.1698371095 -0.0596216843
#> 53   0.055558503 -0.108295716 -0.1530344128  0.0185510752 -0.0017461105
#> 54  -0.013246288  0.068691310 -0.0508117942 -0.0741612859  0.0138410242
#> 55   0.067451820  0.104022250 -0.0585198394 -0.1143532300 -0.0916484932
#> 56   0.019092140 -0.080429053 -0.1043212402 -0.2284301126 -0.0140809391
#> 57   0.029545822 -0.061179395  0.0098753108 -0.0816247162  0.0116497665
#> 58  -0.066932103 -0.075999741 -0.0748840302 -0.0204669799  0.0612926729
#> 59   0.205421758 -0.011054016 -0.0543286236  0.2268260385 -0.0163320222
#> 60   0.164359062 -0.166899630 -0.0232146112  0.0449414152 -0.0345448423
#> 61  -0.002771597  0.079064285 -0.0497118221 -0.0280498885 -0.0646213525
#> 62  -0.041964797  0.116043872  0.0167270296 -0.0658496994  0.0181159518
#> 63  -0.132586839  0.148416293 -0.0640853573 -0.1132884093  0.0778222400
#> 64   0.025802136 -0.034876851  0.0339968700 -0.1857656895  0.0682606996
#> 65  -0.043036565  0.035714698  0.0522672780 -0.0673415885 -0.0325067071
#> 66   0.058109383  0.016023652 -0.0005220509 -0.0538778721 -0.0425244002
#> 67   0.036949672  0.036993688  0.0845272343 -0.0882971074  0.0141268457
#> 68  -0.071956962  0.154672886 -0.1944018747 -0.0466492000 -0.0096734031
#> 69  -0.058241398 -0.178689004  0.1095632363  0.0750484913  0.0081980517
#> 70  -0.008098559  0.075074665 -0.1243735653 -0.1641193445  0.0982428151
#> 71  -0.046161627 -0.083962032  0.0989254724  0.1349589720  0.0138538242
#> 72  -0.030176172  0.009791774  0.1433511740  0.0722414270  0.0578249297
#> 73   0.006323776  0.013685338 -0.0981688278 -0.0683957173  0.0446918860
#> 74  -0.018887249  0.189570368 -0.0769974371  0.0434275180  0.1130387729
#> 75  -0.027122593 -0.088557526  0.1090168529 -0.0824285966 -0.0010881028
#> 76  -0.193910312  0.154624709 -0.0137178863 -0.0488255388  0.0214739788
#> 77   0.013851766  0.134094913 -0.1264986260  0.0240068127 -0.0421871841
#> 78   0.038051521  0.117026422 -0.2150417197  0.0842916131 -0.1446474154
#> 79   0.001460132  0.104360013 -0.0778015591 -0.0398909989 -0.0256686630
#> 80  -0.082081340 -0.103434688 -0.0649564286  0.0898450930  0.0691347891
#> 81  -0.269441254  0.027145381 -0.0541755822  0.0795334136 -0.0989373240
#> 82  -0.031489443 -0.061514311 -0.0128266774  0.0430665752 -0.0057664445
#> 83  -0.095376387 -0.037948289  0.0576039287 -0.0305630302 -0.0202400556
#> 84   0.144241581  0.147374411  0.1570338546 -0.0339163778 -0.0155958882
#> 85   0.031514873 -0.055295313 -0.0927366892  0.0503686264 -0.0155562879
#> 86  -0.033191884  0.021921244  0.0063925003  0.0797402866  0.0405926489
#> 87  -0.092035375 -0.084601403 -0.0445232567  0.1480562172 -0.0261559895
#> 88   0.017590818 -0.016732166 -0.1183124185  0.0097741570 -0.0629246415
#> 89  -0.129488421  0.158580602  0.2346477273  0.0009011458 -0.0634953406
#> 90  -0.056477686 -0.195594339 -0.0175370630  0.0139258453 -0.0002617328
#> 91  -0.134428674  0.046676809 -0.1436609058 -0.0216245189 -0.0922280552
#> 92  -0.116500811 -0.211760382 -0.1253658960 -0.0362212334  0.0469411120
#> 93   0.015228779 -0.095793123  0.0689746538 -0.1577806813 -0.0127092815
#> 94  -0.055718983  0.022826649 -0.0338033159 -0.0162961748 -0.0028743983
#> 95   0.081759530 -0.220183197 -0.1061417772  0.1292090308 -0.0284120763
#> 96  -0.108107944  0.096716706 -0.0353122053  0.2073641601 -0.0215561687
#> 97  -0.051733685 -0.121682657 -0.1444323507 -0.0877338379  0.0478351991
#> 98  -0.087267249 -0.097125187 -0.1301702938 -0.0884856050  0.0080907126
#> 99   0.018037947  0.086616721  0.0405463528 -0.0866438340 -0.0279770004
#> 100 -0.054846616 -0.012143678  0.0703854436  0.1872894944  0.0191628759
#> 
#> $variance
#>          PC1          PC2          PC3          PC4          PC5          PC6 
#> 1.804157e+01 1.443630e+01 1.352075e+01 1.254409e+01 1.098393e+01 9.180965e+00 
#>          PC7          PC8          PC9         PC10 
#> 7.860209e+00 7.128041e+00 6.304136e+00 5.779276e-31 
#> 
#> $sdev
#>  [1] 4.192682e+00 3.750444e+00 3.629570e+00 3.496024e+00 3.271399e+00
#>  [6] 2.990879e+00 2.767399e+00 2.635359e+00 2.478378e+00 7.503976e-16
#> 
#> $metadata
#> NULL
#> 
#> $xvars
#> NULL
#> 
#> $yvars
#> NULL
#> 
#> $components
#>  [1] "PC1"  "PC2"  "PC3"  "PC4"  "PC5"  "PC6"  "PC7"  "PC8"  "PC9"  "PC10"
#> 
#> attr(,"class")
#> [1] "pca"