PCA Analysis

This function performs principal component analysis (PCA) on a given matrix.

Usage

prot.pca(
  mat,
  metadata = NULL,
  center = TRUE,
  scale = FALSE,
  rank = NULL,
  removeVar = NULL,
  transposed = FALSE,
  BSPARAM = BiocSingular::ExactParam(),
  in_workflow = FALSE
)

Arguments

mat

A numeric matrix.

metadata

An optional data frame with rownames matching 'colnames(mat)'.

center

A logical indicating whether to center the data before PCA.

scale

A logical indicating whether to scale the data before PCA.

rank

An integer indicating the number of principal components to calculate.

removeVar

A numeric value between 0 and 1 indicating the fraction of variables to remove based on variance.

transposed

A logical indicating whether the matrix is transposed.

BSPARAM

An object of class ExactParam or ApproxParam.

in_workflow

Indicated whether this function is called from within prot.workflow.

Value

A list with components rotated, loadings, variance, sdev, metadata, xvars, yvars, and components.

References

BiocSingular (https://bioconductor.org/packages/release/bioc/html/BiocSingular.html)

Examples

mat <- matrix(rnorm(1000), nrow = 100, ncol = 10)
prot.pca(mat)
#> $rotated
#>           PC1        PC2       PC3        PC4        PC5        PC6        PC7
#> 1  -4.5749420 -4.9166791 -2.438923 -4.2392888 -0.7218931  0.4682044 -0.5858324
#> 2   1.2126548  2.3927033 -4.175196  3.5685315 -2.1563216 -5.6157992  2.5094790
#> 3  -3.9584788  5.6239816 -4.116374 -1.7628811 -1.9153544  3.1502179  0.8654230
#> 4   1.1613757 -5.3916679  1.493888  0.8608373 -5.8413396  0.7961403 -0.8125442
#> 5  -1.3769439  4.7894275  6.665701 -1.4623597 -0.1159239  0.9054339  2.6818892
#> 6   0.5013518 -0.9099626 -2.514002  5.3640429  4.0767328  4.5709192 -0.8975764
#> 7   5.1972135  2.5120488  2.193003  0.9582878 -2.2465955  1.0785953 -4.0758461
#> 8   5.9854724  0.4028231 -2.038807 -5.5289235  3.5982729 -1.6153261 -1.1306517
#> 9  -5.7919270 -0.2846590  2.822318  1.6105095  2.9990950 -3.9889082 -3.8141372
#> 10  1.6442235 -4.2180157  2.108391  0.6312440  2.3233274  0.2505225  5.2597970
#>           PC8        PC9         PC10
#> 1  -4.5707436  0.1109678 7.907711e-16
#> 2  -1.6331864  0.9871545 7.907711e-16
#> 3   2.9398519 -2.1049818 7.907711e-16
#> 4   3.3820662  2.2228148 7.907711e-16
#> 5  -1.5598596  3.1391561 7.907711e-16
#> 6  -0.7298716  2.4142422 7.907711e-16
#> 7  -2.7099892 -3.4769761 7.907711e-16
#> 8   1.9759705  1.7270375 7.907711e-16
#> 9   2.0875407 -1.0079266 7.907711e-16
#> 10  0.8182212 -4.0114884 7.907711e-16
#> 
#> $loadings
#>              PC1           PC2          PC3           PC4          PC5
#> 1    0.119435690 -0.0597267265  0.171397764  0.0080440785 -0.096282733
#> 2    0.118679478 -0.2212291606  0.030553961  0.0721514971  0.068990174
#> 3   -0.015521986  0.1167298827 -0.255573211  0.0541781813  0.107893603
#> 4    0.076402941 -0.0174448308  0.024827664 -0.1664794870  0.142417930
#> 5   -0.164198979 -0.1116286920 -0.065850188  0.0682081806  0.353248864
#> 6   -0.272514860 -0.1239830352 -0.062902552  0.0348859116 -0.069154879
#> 7    0.062483420  0.0897965935  0.005075963  0.0396889208  0.075908182
#> 8   -0.123149099  0.0067171451  0.059823326 -0.0055125491 -0.007600486
#> 9    0.113756344  0.1329855664 -0.006281167  0.0286263063 -0.027767922
#> 10   0.069542560  0.0349986038  0.012719106  0.1491557425 -0.200485897
#> 11  -0.021923750 -0.0558653839 -0.106246785 -0.0484100504  0.013949204
#> 12   0.030976508  0.0334774233  0.095881636 -0.0774745045 -0.006405546
#> 13  -0.098284466  0.0510884233 -0.068686877  0.2169699858 -0.095141081
#> 14  -0.030839492  0.0487909607  0.043225617  0.0167680571 -0.046550728
#> 15  -0.158372404 -0.1003261356  0.035669684  0.0969161182  0.014140655
#> 16   0.078006882  0.0180687933  0.095299780  0.1500819303 -0.028987118
#> 17   0.104627848 -0.0814457616 -0.098044398 -0.0392996102  0.064037152
#> 18   0.017873145 -0.3046202629 -0.202543969  0.0104480714 -0.113705451
#> 19  -0.084267165  0.1401146101 -0.048738742 -0.0919442719  0.028438478
#> 20  -0.006209130  0.0212958915 -0.030339490  0.0752488080  0.008978580
#> 21   0.038087090 -0.0694604499 -0.005323471 -0.1205955061 -0.219330246
#> 22   0.033766190  0.1101005682 -0.060770740 -0.0865879783 -0.182246009
#> 23   0.128887199  0.0664441152  0.139693764  0.0185937827 -0.073908731
#> 24  -0.091628129 -0.0773538983  0.093601310 -0.0940101009 -0.051051220
#> 25  -0.069871241 -0.0343396138 -0.059382144  0.1230665226  0.187609441
#> 26  -0.152667480 -0.0257986901  0.050629387  0.0453270834 -0.140792177
#> 27   0.050712331 -0.0864281723  0.091791354 -0.1378031550  0.041176960
#> 28   0.136150572 -0.0959881788 -0.004539145 -0.0807107984  0.112465211
#> 29   0.019669544 -0.1308812044 -0.024463419  0.0587019785  0.043253983
#> 30   0.089592178  0.0416492043 -0.070244979  0.0220493537  0.171450559
#> 31  -0.040934047  0.0366141786 -0.079954641 -0.0759856918  0.062109428
#> 32  -0.166879713  0.0271023275  0.021639214  0.1354718387  0.031435946
#> 33   0.037319592 -0.2111121423 -0.038543102  0.1527649857 -0.033553609
#> 34   0.054713023 -0.0531331357 -0.090059684  0.1473484814  0.042711295
#> 35  -0.070622131  0.0008185636  0.087661284  0.0216586508 -0.146651247
#> 36   0.089905959  0.0614528334 -0.251711873  0.1232061321 -0.095810622
#> 37   0.084950465  0.1392031473 -0.201805214 -0.1617678893  0.017103463
#> 38  -0.011327457  0.0123344997  0.073923014  0.0881976382  0.024343760
#> 39  -0.135379443 -0.1069179519 -0.107664109 -0.0872917279 -0.189499441
#> 40   0.054321996 -0.0937731632 -0.017036492 -0.0477538353  0.012354240
#> 41  -0.105700802 -0.0626605713 -0.077141552  0.1010865851  0.057418088
#> 42  -0.098376925 -0.0110225838  0.119234372 -0.0698740075  0.159452427
#> 43   0.143782255  0.1062684214 -0.007834187  0.0716232587 -0.162889325
#> 44  -0.068812136  0.0265252670  0.097044896 -0.0864536392  0.025669796
#> 45  -0.047708967 -0.0027955394  0.093619297  0.0257661266 -0.028986767
#> 46   0.131298992 -0.1590350484  0.012473714  0.0609970317  0.153403709
#> 47   0.032484709 -0.2249384988  0.004257894  0.0977390879 -0.052355495
#> 48   0.030506773  0.0057789675 -0.091566113  0.0004787558 -0.110034490
#> 49   0.129294060 -0.1728199570 -0.094110451  0.0004288259 -0.162884784
#> 50   0.116578666  0.0362315496  0.025784066  0.0769963847  0.075452032
#> 51  -0.249875496 -0.2312531842  0.040664796 -0.0075590499 -0.098436340
#> 52   0.094297552  0.0566928968  0.013112756  0.1013659927  0.074835509
#> 53  -0.067407219  0.0977848372 -0.053134992  0.0109947385 -0.048446836
#> 54  -0.105715270 -0.0109318513  0.033585489 -0.1411539128  0.096984595
#> 55   0.205856293 -0.0430337499  0.010732075  0.0130993950 -0.053822428
#> 56   0.066112692 -0.0633796107  0.138917570 -0.1932942361 -0.120012673
#> 57   0.024373255 -0.0957229903 -0.013450845  0.2160751868 -0.139373375
#> 58   0.101997224 -0.0088496467  0.065068735 -0.1594250118 -0.117153573
#> 59   0.066601000 -0.0847279899  0.094845628 -0.0379918033 -0.025986856
#> 60  -0.086381432  0.1203534580 -0.159377621 -0.2209464937  0.016957642
#> 61   0.098943276  0.0004790741 -0.017831221 -0.0794224124  0.034344049
#> 62   0.038969075 -0.0580725451  0.093760298 -0.0334001679  0.005953101
#> 63   0.081854010  0.1415102283  0.137617694  0.1133503229 -0.025641999
#> 64  -0.148509219  0.0155757265  0.005307912  0.1060475449 -0.054860625
#> 65  -0.002119053 -0.1236072679  0.133143785 -0.0010623086  0.034696899
#> 66   0.024543944  0.0058753116 -0.177919358 -0.1204887020 -0.026179888
#> 67   0.037665114 -0.0422169179 -0.197362903 -0.0341762918  0.042806224
#> 68   0.003860103 -0.0839646702 -0.168107622  0.0295030708  0.013399056
#> 69  -0.002301441  0.0723304558 -0.179834364 -0.0458693050 -0.147745844
#> 70  -0.162656643  0.0540675933 -0.167587701  0.0658736219 -0.048867486
#> 71   0.054927732  0.0405506365  0.033596308 -0.1153811646 -0.025736736
#> 72  -0.029525564 -0.0045914006  0.024712716 -0.0145583869 -0.002738279
#> 73   0.065409089 -0.1127100104 -0.166262653  0.0213572711 -0.063993164
#> 74  -0.228589682 -0.0822383367 -0.036416950 -0.1264842804 -0.012194861
#> 75  -0.127176744 -0.1351257228  0.074141586 -0.0828201679  0.116192217
#> 76  -0.125639832 -0.0093322414  0.125773212 -0.2246249863 -0.068170810
#> 77  -0.107150567  0.0703462818 -0.013148416  0.0650850430 -0.053188916
#> 78   0.128802338  0.0543264472  0.069566634  0.0108678983 -0.038479777
#> 79   0.033711331  0.0801336044 -0.183714436  0.0842991747 -0.064660311
#> 80   0.241391953  0.1048413144  0.133547165  0.1395467396  0.026928957
#> 81  -0.119018559  0.0068570797 -0.034486485  0.2775327383 -0.119992487
#> 82  -0.081036815  0.1999024219 -0.020866898 -0.0689626760 -0.042143690
#> 83   0.141585844 -0.1141375946 -0.161139304 -0.0224573288  0.020702063
#> 84   0.036194756 -0.0218629062 -0.106492730  0.0229556313  0.156354995
#> 85  -0.033902421 -0.1282867541  0.004971481 -0.1081284508  0.036817931
#> 86   0.064535711 -0.1368565906  0.018762568 -0.0769756533 -0.113888854
#> 87  -0.014275786 -0.0061860859 -0.076533690 -0.0907882129  0.101644256
#> 88  -0.028266747  0.0717406390 -0.229945678 -0.1233361758 -0.064458724
#> 89   0.076122209 -0.2465168953 -0.006259549 -0.0662622903 -0.109900312
#> 90   0.131969119 -0.0617995917 -0.126109921 -0.1309626145  0.186381902
#> 91   0.047359054 -0.0278281429 -0.073077540 -0.0202507655 -0.144656524
#> 92   0.062985887  0.0090469974 -0.167539355 -0.0623709432  0.008698742
#> 93  -0.075500296  0.1997690236  0.011000968 -0.1128801829 -0.169395829
#> 94   0.078590633 -0.0736813296  0.030310595 -0.0108167379 -0.003630905
#> 95  -0.047108697  0.0548398239  0.056521926  0.1624928832  0.119234835
#> 96  -0.046819527  0.1127508122  0.042847295  0.0906748836  0.085647883
#> 97   0.030923189  0.0674614055 -0.076641060  0.0575349008  0.006871830
#> 98   0.015187802  0.0794467356 -0.003223348 -0.0733851306  0.128264962
#> 99  -0.075436618  0.0571604456  0.048618580  0.0198105310  0.038153101
#> 100 -0.028015843  0.0168249211  0.082092163 -0.0819824970 -0.069592340
#>              PC6          PC7           PC8           PC9          PC10
#> 1    0.172015306  0.155674621 -0.0045350496  0.1491235101  0.1924624685
#> 2   -0.069145942  0.031578500  0.1155893776  0.1508719640  0.0614225262
#> 3    0.057068938  0.037036687  0.0747176037 -0.2295634175  0.4102550457
#> 4   -0.181992242  0.016170744  0.1129877407  0.1006329017  0.2593467253
#> 5    0.101280732 -0.026641505  0.0647224425  0.1212197231  0.2239133249
#> 6   -0.034668781  0.028906566 -0.0194674308 -0.0537007182 -0.2942939934
#> 7   -0.029201754 -0.025720159 -0.0541593580  0.0457233319 -0.2036770117
#> 8   -0.059100803  0.112158463  0.0343072505  0.0551320342 -0.1986887914
#> 9   -0.218577192  0.104870595  0.1255123626  0.0708370221 -0.2045562886
#> 10   0.185957519  0.064209216 -0.0695312080 -0.0052967970  0.1103779299
#> 11   0.150339576 -0.056394260  0.0822647701 -0.0403713434  0.0011475803
#> 12  -0.118927490 -0.012066023 -0.2214443835  0.0101809401  0.1249482639
#> 13   0.092885844  0.208351049  0.0298574935  0.0742108566  0.0576051831
#> 14  -0.307113123 -0.188175628 -0.0180453001  0.0427070287 -0.0319736097
#> 15  -0.091703963  0.124705878  0.0613606575 -0.0710448390 -0.0538089966
#> 16  -0.103710237  0.010673077  0.1429939983 -0.0218854609  0.0607863173
#> 17   0.004449391  0.028440690 -0.0169148406 -0.1189707369  0.0053077719
#> 18   0.014610801 -0.070852986  0.0008039563  0.0590221030  0.0920437221
#> 19   0.012423032  0.107103715 -0.0066555388  0.0239217860 -0.0445736870
#> 20  -0.097498218 -0.076512970  0.1542486562 -0.0194655778  0.0232994490
#> 21  -0.184227523 -0.227233414  0.1146478165 -0.1095152321  0.1997072607
#> 22   0.101711553 -0.022176904  0.0504988038  0.1672900080 -0.0283200239
#> 23  -0.110259989 -0.124488451  0.0155473340  0.1519384937  0.0420477478
#> 24   0.031171879 -0.042727673  0.0591864588 -0.0165526691  0.0138123650
#> 25  -0.132023236  0.049866404 -0.1381448180 -0.0246090981 -0.0091396648
#> 26   0.006996446  0.073339937  0.0803754718  0.0247410722  0.1187227148
#> 27   0.116474652 -0.001625594 -0.2372751741  0.1308214256  0.1586820757
#> 28   0.079976141  0.005766531  0.0847531922 -0.1597198227 -0.0319985199
#> 29  -0.052779879  0.061603278  0.1335187949  0.1008708082 -0.0149383123
#> 30   0.010938853 -0.078356212  0.1294267557  0.1285499015 -0.1641431478
#> 31   0.001955698  0.114813863  0.0641912579 -0.0306736806 -0.0109347850
#> 32  -0.101746896  0.029287758  0.1331774129 -0.1139440149  0.0170707794
#> 33   0.030946308 -0.091632795 -0.1520338855 -0.0709783163 -0.0140826582
#> 34   0.097625296 -0.157129743 -0.1185030683  0.0249463079  0.0953512960
#> 35   0.110347682 -0.013373263  0.2278452227  0.0350843614  0.0089367084
#> 36   0.077043662  0.046571103 -0.0879275721  0.0038548603  0.0060599545
#> 37  -0.009701670  0.102742751  0.0046404459  0.0631115806  0.0297085104
#> 38  -0.127701195 -0.004314883 -0.2466066386 -0.0615829142  0.0718579710
#> 39  -0.078337692  0.009645018 -0.0362557447 -0.0522356197  0.0340569786
#> 40   0.082911985  0.044003254  0.0376743086 -0.2016178611 -0.0287408229
#> 41  -0.050031043 -0.002489708 -0.0290726959 -0.0518930671 -0.0419043186
#> 42   0.101586519  0.078397408  0.1095425922  0.0656829204  0.0006329643
#> 43  -0.069101639  0.099123050  0.0837903698  0.0635760517  0.0817665463
#> 44  -0.112787669  0.078590931 -0.1362679419  0.1399219682  0.0231207634
#> 45  -0.086151752  0.048840759 -0.0146908101 -0.0576176379  0.0553974904
#> 46   0.129427308 -0.025293020  0.0406840789 -0.0202218568 -0.0504403204
#> 47   0.056109280  0.053817841  0.1082909565 -0.1086537935 -0.0218422136
#> 48   0.015075008  0.060152042  0.1283077236  0.0139623109 -0.0948259054
#> 49  -0.063487205 -0.006927638 -0.0079771495 -0.1130064176 -0.0059304585
#> 50   0.047049666  0.254613735  0.1758343100 -0.1066932126  0.1363278312
#> 51  -0.066270568 -0.085144143 -0.0291810562  0.0365920625  0.0955326516
#> 52   0.066517468  0.079520475 -0.1431915117  0.1161513884 -0.0851776261
#> 53  -0.041276825  0.017587609 -0.0372535052 -0.2023468192 -0.0023275815
#> 54   0.155307776 -0.042722839 -0.0227188549  0.1834570907 -0.0245094101
#> 55  -0.102784918 -0.042429430 -0.0005402844  0.0807324668  0.0375899329
#> 56   0.035202104  0.248698373 -0.0798180620 -0.0973250262 -0.0042797109
#> 57   0.119384342 -0.043885555  0.0064383847  0.0005545124 -0.0044206145
#> 58   0.004087395 -0.025142747  0.0722934543 -0.0669323228 -0.0883254910
#> 59  -0.105197563 -0.087402155  0.1044652082  0.1930160760  0.0582780825
#> 60   0.109198966 -0.088543077 -0.0450457902 -0.0392333044  0.0642063643
#> 61  -0.078465659  0.041582103 -0.1068711747  0.0023988309  0.0144202852
#> 62   0.062712534  0.053610344  0.0120756533  0.0418465163 -0.0852216844
#> 63  -0.070028511 -0.108040717  0.0331950313 -0.1860056601  0.0459053328
#> 64  -0.080968322  0.028785170  0.1954096775  0.0687919815  0.0245924882
#> 65  -0.102589632 -0.052563958 -0.1231846605 -0.1744106704  0.1515335144
#> 66  -0.117715455  0.022916390 -0.1609922321  0.0324704440  0.0618551296
#> 67  -0.119180600  0.055482345  0.0064985279 -0.1319280710 -0.0636822157
#> 68   0.019444096  0.026862070 -0.1295486922 -0.0519182843  0.0178232618
#> 69   0.017207781 -0.122734076  0.0774835246  0.0999750485  0.0867127807
#> 70   0.209281215 -0.134849002  0.0618757266  0.1130520882  0.0866006202
#> 71   0.102166765 -0.026051791  0.0027472377 -0.1863272230 -0.0082511908
#> 72   0.088951904  0.159063490  0.0056520418 -0.1009843304  0.0295815528
#> 73  -0.196894799  0.176789451  0.1301193561  0.0830771047 -0.0465575127
#> 74   0.023714691 -0.043699432  0.0467342067  0.0082036618  0.0223549051
#> 75   0.024365758  0.144414837 -0.0480964804  0.0723359146 -0.0431074811
#> 76  -0.003676970  0.053364309  0.0505823288 -0.1032534571  0.0950244496
#> 77  -0.110415777 -0.021967887 -0.0783346197  0.1080972597  0.0065013463
#> 78   0.082791320  0.179041309 -0.0303192755 -0.0237297759 -0.0027198894
#> 79  -0.016069524 -0.185931135 -0.1584786234 -0.0284111326 -0.1646591641
#> 80   0.054448855 -0.079125171  0.0286260026 -0.0366945105  0.1125857795
#> 81  -0.066551821  0.182257827 -0.0416928105  0.1293297144  0.0837542410
#> 82  -0.029414714  0.002305546 -0.0795647686  0.1430636093  0.0667405443
#> 83  -0.127284103  0.132555303  0.1192300065 -0.0801264731 -0.0307620128
#> 84  -0.037269105  0.078378135 -0.0234067195  0.1250071403 -0.0125249441
#> 85  -0.058431202 -0.016802675  0.0861777176  0.0260913815  0.0605996746
#> 86   0.045542237 -0.169039463 -0.0255627762  0.1146802596 -0.0400728192
#> 87  -0.065681742 -0.108483805  0.1180631140 -0.1161860260 -0.0737469193
#> 88  -0.121582339  0.088569567  0.0914591892  0.1205332343  0.1811623275
#> 89   0.013071187  0.121263206 -0.0764724774  0.1131069423 -0.0890738460
#> 90  -0.136411739  0.159292209  0.0267710390  0.1005335145  0.0081149502
#> 91   0.056272503  0.141559481 -0.0526953844  0.0050111089 -0.0797364132
#> 92  -0.001671024  0.001546825 -0.1158965211  0.0317681645 -0.0635348981
#> 93  -0.026291570  0.108083071  0.0322497686 -0.0528124748 -0.0117300456
#> 94  -0.049922768 -0.224582681  0.1821260501  0.0340558473 -0.0633862279
#> 95  -0.179399118 -0.008765519 -0.0412484256  0.0855842239  0.0963004869
#> 96   0.019815798 -0.007122569  0.0699034027 -0.1606380559 -0.0156656618
#> 97   0.164681664 -0.110202422  0.1522936505  0.1992088041 -0.0309252783
#> 98   0.098232143 -0.143300148  0.1658844371 -0.0613289016 -0.0981504734
#> 99  -0.127162137  0.007509622  0.0031570948 -0.0535205396  0.0101192839
#> 100  0.071475396  0.069023590  0.1338827335  0.0113336130 -0.0160262325
#> 
#> $variance
#>          PC1          PC2          PC3          PC4          PC5          PC6 
#> 1.622288e+01 1.599876e+01 1.322777e+01 1.152771e+01 1.060906e+01 9.662467e+00 
#>          PC7          PC8          PC9         PC10 
#> 8.783349e+00 7.252762e+00 6.715245e+00 7.212405e-31 
#> 
#> $sdev
#>  [1] 3.953242e+00 3.925840e+00 3.569709e+00 3.332429e+00 3.196892e+00
#>  [6] 3.050939e+00 2.908838e+00 2.643268e+00 2.543433e+00 8.335459e-16
#> 
#> $metadata
#> NULL
#> 
#> $xvars
#> NULL
#> 
#> $yvars
#> NULL
#> 
#> $components
#>  [1] "PC1"  "PC2"  "PC3"  "PC4"  "PC5"  "PC6"  "PC7"  "PC8"  "PC9"  "PC10"
#> 
#> attr(,"class")
#> [1] "pca"