Reads flux data from a file or data frame

Reads flux data from a file or data frame

Usage

read_flux(
  file.data = dat_flux,
  csvsep = ";",
  dec = ".",
  sheet.data = 1,
  file.match = "eq_bigg.csv",
  sheet.match = 1,
  rename.reaction = TRUE,
  rescale.reaction = NULL,
  col.id = "Reaction ID",
  col.eq = "Equation",
  col.flux = "Value (%)",
  col.sd = "Standard Deviation",
  FBA = FALSE
)

Arguments

file.data

Path to the file containing the flux data. Alternatively, a data frame can be provided.

csvsep

Separator used in the file. Default is ";".

dec

Decimal separator used in the file. Default is ".".

sheet.data

Sheet number to read from the file. Default is 1.

file.match

Path to the file containing the equation-BiGG name matches.

sheet.match

Sheet number to read from the file matching equations to BiGG reaction names. Default is 1.

rename.reaction

Logical, whether to rename reactions according to BiGG standard. Default is TRUE.

rescale.reaction

Name of a reaction to normalize all fluxes (will be set as 100%). Default is NULL.

col.id

Name of the column containing reaction IDs. Default is "Reaction ID".

col.eq

Name of the column containing carbon transition equations. Default is "Equation".

col.flux

Name of the column containing flux values. Default is "Value (%)".

col.sd

Name of the column containing standard deviation values. Default is "Standard Deviation".

FBA

Logical, whether the data is from an FBA simulation. Default is FALSE.

Value

A data frame with reaction IDs, equations, flux values, and standard deviation.