Filter genes based on missing values

rna.filter_missing filters a transcriptomics dataset based on missing values. Different types of filtering can be applied, which range from only keeping genes without missing values to keeping genes with a certain percent valid values in all samples or keeping genes that are complete in at least one condition.

Usage

rna.filter_missing(
  se,
  type = c("complete", "condition", "fraction", NULL),
  thr = NULL,
  min = NULL
)

Arguments

se

SummarizedExperiment object, transcriptomics data parsed with rna.read_data.

type

(Character string) "complete", "condition" or "fraction", Sets the type of filtering applied. "complete" will only keep genes with valid values in all samples. "condition" will keep genes that have a maximum of thr missing values in at least one condition. "fraction" will keep genes that have a min fraction of valid values in all samples.

thr

(Integer) Sets the threshold for the allowed number of missing values in at least one condition if type = "condition". In other words: "keep genes that have a maximum of 'thr' missing values in at least one condition."

min

(Numeric) Sets the threshold for the minimum fraction of valid values allowed for any gene if type = "fraction".

Value

A filtered SummarizedExperiment object.