Plot bar plots for gene abundance or fold change

This function generates bar plots for gene abundance or fold change, normalized by either a reference gene or centered to the mean. The user can also plot log2 intensities, contrasts, or raw counts.

Usage

rna.plot_bar(
  dds,
  genes,
  type = c("contrast", "log_centered", "centered", "raw", "log2"),
  shrunken.lfc = TRUE,
  contrast = NULL,
  conditions = NULL,
  col.id = "ID",
  name_table = NULL,
  match.id = "Accession",
  match.name = "Name",
  convert_name = FALSE,
  shape.size = 2.5,
  collage = TRUE,
  plot = TRUE,
  export = TRUE,
  export.nm = NULL
)

Arguments

dds

A DESeqDataSet object containing the data.

genes

A vector of gene names or IDs.

type

Type of plot: "contrast", "log_centered", "centered", "raw", or "log2".

"log_centered": log2-transformed intensities centered to their mean.

"centered": raw intensities centered to their mean.

"contrast": log2 fold changes for specified contrasts.

"raw": raw (or normalized) read counts with mean and 95% CI.

"log2": log2-transformed intensities with mean and 95% CI.

shrunken.lfc

A logical indicating if shrunken log2 fold changes should be used if type = "contrast". Default is TRUE.

contrast

Name of the contrast for plotting when type is "contrast".

conditions

Character vector of condition names to be plotted.

col.id

Name of the column with gene IDs in dds.

name_table

A data frame containing the mapping of gene names to reference names.

match.id

The name of a column in name_table to find matches with col.id.

match.name

The name of a column in name_table to assign new names based on matches of col.id and match.id.

convert_name

Logical indicating whether to convert gene names based on name_table.

shape.size

Numeric; replicate symbol size in the plot.

collage

Logical; combine multiple gene plots into collages of up to 8. Default TRUE.

plot

Logical; print the plot if TRUE or return data frame if FALSE. Default TRUE.

export

Logical; export plots as PDF/PNG if TRUE. Default TRUE.

export.nm

Base name for exported files if export = TRUE.

Value

If plot = FALSE and length(genes) <= 8, returns a data frame of values and 95% CIs.